https://online.sbis.ru/opendoc.html?guid=8d622a59-3886-4034-935e-b5d7b074ed74
доработка для якоря (фрагмента) url адреса, тест для фрагмента url адреса
